pyproject.toml
[tool.poetry]
name = "actual-project"
version = "0.1.0"
description = ""
authors = ["Andrey Kopylov <104315220+obscurezz@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "actual_project"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
psycopg2 = "^2.9.5"
django-environ = "^0.9.0"
gunicorn = "^20.1.0"
djangorestframework = "^3.14.0"
django-filter = "^22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
